requirements.txt
# FastAPI and web server dependencies
fastapi>=0.115.2
uvicorn[standard]>=0.24.0
python-multipart>=0.0.18

# Authentication and security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.1

# Document processing
PyPDF2>=3.0.1
python-docx>=1.1.0

# HTTP requests for LM Studio API
requests>=2.32.2
httpx>=0.25.2

# Utilities
python-dotenv>=1.0.0

# CORS support
fastapi-cors>=0.0.6

# File handling
aiofiles>=23.2.1

# Data validation
pydantic>=2.7.4

# Additional utilities
typing-extensions>=4.11.0

# Starlette (FastAPI dependency)
starlette>=0.40.0